requirements.txt
# Core ML/AI
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
faiss-gpu>=1.7.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0

# API & Utilities
requests>=2.31.0
tiktoken>=0.5.0
tqdm>=4.65.0
python-dotenv>=1.0.0

# Configuration
pyyaml>=6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Testing
pytest>=7.0.0

# Development
jupyter>=1.0.0
ipykernel>=6.0.0
